refactor(service-providers): design system compliance and import cleanup
ServiceProvidersTable:
- fix barrel imports: Button, cn, formatNumberUserReadable now import
  directly from source files
- bg-orange-400/12 → bg-surface-opacity-brand (design system token)
- after:bg-[var(--color-border-default)] → after:bg-border-default
- text-blue-400/text-pink-500 → text-[var(--base-chart-1/2)] (design
  system chart color tokens)

StatusCell:
- fix barrel import: cn now imports directly from @/shared/utils/cn
- text-[13px] → text-alternative-sm (design system text size token)
